SUGGESTED CHANGES
 - Remove getNodeListByName since we don't have a database dropdown
   May add back later in a later phase

 - Get rid of redundant paramSchema

 - pgt_chain.js change returncode to status
   to match with the name in sql

 - Get rid of redundant el.last_run in chain properties.sql

 - run_now should state it is using notification instead of updating
   tables. Also add in notice which client was notified
   and if no client could be notified. Include tests to test
   new messaging.

 - Fix statistics docstrings copy-paste error.
   Should be talking about chains

 - Fix int/bigint datatype mismatch for chain_id and task_id

BUG FIXES

 - Chain properties tab is not showing the true last run,
   should use finished for sorting
   and also account for running jobs. Also expose status.

 - Fix W503 python style issues

+++ web/pgadmin/browser/server_groups/servers/pg_timetable/templates/pgt_chain/sql/default/create.sql
@@ -1,7 +1,7 @@
 DO $$
 DECLARE
-    cid integer;
-    tid integer;
+    cid bigint;
+    tid bigint;
 BEGIN
 INSERT INTO timetable.chain(
     chain_name, run_at, client_name, live, max_instances, timeout, self_destruct, exclusive_execution, on_error
